backup_requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosmtplib==3.0.2
ale-py==0.10.2
alembic==1.14.1
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
backtesting==0.6.3
bcrypt==3.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
bokeh==3.7.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.2.1
decorator==5.2.1
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.115.11
fastapi-mail==1.4.2
filelock==3.17.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-pasta==0.2.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
ipython==8.34.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jose==1.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.39.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
MetaTrader5==5.0.4874
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.29.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.56.4
numpy==1.26.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pandas-ta==0.3.14b0
parso==0.8.4
passlib==1.7.4
pillow==11.1.0
plotly==6.0.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
python-telegram-bot==21.11.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stable_baselines3==2.5.0
stack-data==0.6.3
starlette==0.46.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==2.1.4
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
